refactor: remove added dead code allowances
