+++ crates/monochange/src/publish_readiness.rs
@@ -109,20 +109,6 @@ pub(crate) fn run_publish_readiness(
 	render_report(&report, options.format)
 }
 
-#[allow(dead_code)]
-pub(crate) fn validate_publish_readiness_artifact(
-	root: &Path,
-	configuration: &WorkspaceConfiguration,
-	prepared_release: Option<&PreparedRelease>,
-	selected_packages: &BTreeSet<String>,
-	artifact_path: &Path,
-) -> MonochangeResult<()> {
-	let artifact = read_report_artifact(artifact_path)?;
-	#[rustfmt::skip]
-	let current_report = build_publish_readiness_report_for_publish(root, configuration, prepared_release, selected_packages)?;
-	validate_publish_readiness_report(&artifact, &current_report)
-}
-
 pub(crate) fn publish_plan_package_filter_from_readiness_artifact(
 	root: &Path,
 	configuration: &WorkspaceConfiguration,
@@ -438,19 +424,6 @@ fn read_report_artifact(input: &Path) -> MonochangeResult<PublishReadinessReport
 	serde_json::from_str(&body).map_err(publish_readiness_json_error)
 }
 
-#[allow(dead_code)]
-fn validate_publish_readiness_report(
-	artifact: &PublishReadinessReport,
-	current: &PublishReadinessReport,
-) -> MonochangeResult<()> {
-	validate_readiness_artifact_header(artifact)?;
-	validate_readiness_artifact_status(artifact)?;
-	validate_readiness_current_status(current)?;
-	validate_readiness_release_commit(artifact, current)?;
-	validate_readiness_input_fingerprint(artifact, current)?;
-	validate_readiness_packages(artifact, current)
-}
-
 fn validate_publish_readiness_plan_artifact(
 	artifact: &PublishReadinessReport,
 	current: &PublishReadinessReport,
@@ -477,26 +450,6 @@ fn validate_readiness_artifact_header(report: &PublishReadinessReport) -> Monoch
 	Ok(())
 }
 
-#[allow(dead_code)]
-fn validate_readiness_artifact_status(report: &PublishReadinessReport) -> MonochangeResult<()> {
-	if report.status == PublishReadinessGlobalStatus::Ready {
-		return Ok(());
-	}
-	Err(MonochangeError::Config(
-		"publish readiness artifact is blocked; rerun `mc publish-readiness` and resolve blockers before `mc publish`".to_string(),
-	))
-}
-
-#[allow(dead_code)]
-fn validate_readiness_current_status(report: &PublishReadinessReport) -> MonochangeResult<()> {
-	if report.status == PublishReadinessGlobalStatus::Ready {
-		return Ok(());
-	}
-	Err(MonochangeError::Config(
-		"current publish readiness is blocked; rerun `mc publish-readiness` and resolve blockers before `mc publish`".to_string(),
-	))
-}
-
 fn validate_readiness_release_commit(
 	artifact: &PublishReadinessReport,
 	current: &PublishReadinessReport,
@@ -507,37 +460,6 @@ fn validate_readiness_release_commit(
 	Err(MonochangeError::Config(format!(
 		"publish readiness artifact was generated for release record {}, but `mc publish` selected {}; rerun `mc publish-readiness --from HEAD --output <PATH>`",
 		artifact.record_commit, current.record_commit
-	)))
-}
-
-#[allow(dead_code)]
-fn validate_readiness_packages(
-	artifact: &PublishReadinessReport,
-	current: &PublishReadinessReport,
-) -> MonochangeResult<()> {
-	let artifact_packages = package_identities(&artifact.packages)?;
-	let current_packages = package_identities(&current.packages)?;
-	if artifact.package_set_fingerprint != package_set_fingerprint(&artifact.packages) {
-		return Err(MonochangeError::Config(
-			"publish readiness artifact package fingerprint does not match its package list"
-				.to_string(),
-		));
-	}
-	if artifact_packages == current_packages {
-		return Ok(());
-	}
-	let missing = current_packages
-		.difference(&artifact_packages)
-		.map(render_package_identity)
-		.collect::<Vec<_>>();
-	let stale = artifact_packages
-		.difference(&current_packages)
-		.map(render_package_identity)
-		.collect::<Vec<_>>();
-	Err(MonochangeError::Config(format!(
-		"publish readiness artifact package set is stale or does not match selected packages (missing: {}; stale: {})",
-		render_package_identity_list(&missing),
-		render_package_identity_list(&stale)
 	)))
 }
