+++ crates/monochange_changelog/src/lib.rs
@@ -1347,48 +1347,6 @@ fn push_unique_release_note_entry(entries: &mut Vec<String>, entry: String) {
 	}
 }
 
-#[allow(dead_code)]
-struct ResolvedSectionDefinition {
-	types: Vec<String>,
-}
-
-#[allow(dead_code)]
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-enum ResolvedReleaseSectionTarget {
-	Section(usize),
-	Uncategorized,
-}
-
-#[allow(dead_code)]
-fn section_matches_resolved_type(section: &ResolvedSectionDefinition, change_type: &str) -> bool {
-	section
-		.types
-		.iter()
-		.any(|candidate| candidate.trim() == change_type)
-}
-
-#[allow(dead_code)]
-fn classify_release_note_change(
-	change: &ReleaseNoteChange,
-	sections: &[ResolvedSectionDefinition],
-) -> ResolvedReleaseSectionTarget {
-	if let Some(change_type) = change.change_type.as_deref()
-		&& let Some(index) = sections
-			.iter()
-			.position(|section| section_matches_resolved_type(section, change_type))
-	{
-		return ResolvedReleaseSectionTarget::Section(index);
-	}
-	let bump_selector = change.bump.to_string();
-	if let Some(index) = sections
-		.iter()
-		.position(|section| section_matches_resolved_type(section, &bump_selector))
-	{
-		return ResolvedReleaseSectionTarget::Section(index);
-	}
-	ResolvedReleaseSectionTarget::Uncategorized
-}
-
 #[cfg(test)]
 #[path = "__tests__/changelog_tests.rs"]
 mod tests;
