+++ crates/monochange_dart/src/lib.rs
@@ -706,13 +706,6 @@ pub fn default_dependency_version_prefix() -> &'static str {
 pub fn default_dependency_fields() -> &'static [&'static str] {
 	&["dependencies", "dev_dependencies"]
 }
-#[allow(dead_code)]
-fn yaml_bool(mapping: &Mapping, key: &str) -> Option<bool> {
-	mapping
-		.get(Value::String(key.to_string()))
-		.and_then(Value::as_bool)
-}
-
 fn yaml_mapping<'map>(mapping: &'map Mapping, key: &str) -> Option<&'map Mapping> {
 	mapping
 		.get(Value::String(key.to_string()))
