+++ crates/monochange_gitea/src/__tests__/lib_tests.rs
@@ -39,6 +39,10 @@ use monochange_test_helpers::git;
 use monochange_test_helpers::git_output;
 use tempfile::tempdir;
 
+fn gitea_client() -> MonochangeResult<Client> {
+	build_http_client("Gitea")
+}
+
 use super::*;
 
 fn must_ok<T, E: std::fmt::Display>(result: Result<T, E>, context: &str) -> T {
